+++ src/main/kotlin/br/com/gamemods/j2nwc/internal/chunk-converter.kt
@@ -43,9 +43,7 @@ internal fun Chunk.toNukkit(
         javaChunk.position,
         toNukkitBiomes(javaChunk.biomes),
         byteArrayOf(0, 0, 0, 0),
-        ByteArray(256) {
-            255.toByte()
-        }
+        IntArray(256) { 255 }
     )
     return nukkitChunk
 }
